update share on social flow/design/text
